[OV-29] refactor : 엔티티 패키지 생성 및 엔티티 이동
- 도메인 아래 각 패키지 생성후 도메인 이동하였습니다
